// frontend/src/app/models/financial-management/financeImportLogData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * Custom financeImportLog
 */
export interface FinanceImportLogData {
    /** custom courseCode */
    courseCode?: string;
    /** coursepublicId */
    coursepublicId?: number;
    /** custom วันที่นำเข้า */
    createDate?: string;
    /** custom วันที่ชำระเงินใช้ในการค้นหา */
    createDateList?: string[];
    /** failRow */
    failRow?: number;
    /** fileReferenceCode */
    fileReferenceCode?: string;
    first?: number;
    /** impFileLink */
    impFileLink?: string;
    /** impFileMoney */
    impFileMoney?: number;
    /** impFileName */
    impFileName?: string;
    /** impFileRow */
    impFileRow?: number;
    /** impFileSize */
    impFileSize?: number;
    /** PK */
    impId?: number;
    /** messageError */
    messageError?: string;
    /** missRow */
    missRow?: number;
    mode?: string;
    /** passRow */
    passRow?: number;
    /** custom publicNameEn */
    publicNameEn?: string;
    /** custom publicNameTh */
    publicNameTh?: string;
    /** Output file zip cert, receipt */
    resultFileZip?: string;
    rowNum?: number;
    size?: number;
    /** custom prcessing, imported, generated, success */
    status?: string;
}